--- src/irx/arrow.py
@@ -1,7 +1,0 @@
-"""
-title: Compatibility exports for legacy Arrow AST nodes.
-"""
-
-from irx.array import ArrayInt32ArrayLength, ArrowInt32ArrayLength
-
-__all__ = ["ArrayInt32ArrayLength", "ArrowInt32ArrayLength"]
