+++ src/irx/builder/lowering/__init__.py
@@ -2,10 +2,7 @@
 title: Concern-grouped lowering mixins.
 """
 
-from irx.builder.lowering.array import (
-    ArrayVisitorMixin,
-    ArrowVisitorMixin,
-)
+from irx.builder.lowering.array import ArrayVisitorMixin
 from irx.builder.lowering.binary_ops import BinaryOpVisitorMixin
 from irx.builder.lowering.buffer import BufferVisitorMixin
 from irx.builder.lowering.control_flow import (
@@ -21,7 +18,6 @@ from irx.builder.lowering.variables import VariableVisitorMixin
 
 __all__ = [
     "ArrayVisitorMixin",
-    "ArrowVisitorMixin",
     "BinaryOpVisitorMixin",
     "BufferVisitorMixin",
     "ControlFlowVisitorMixin",
