--- src/irx/builder/lowering/arrow.py
@@ -1,7 +1,0 @@
-"""
-title: Compatibility exports for legacy Arrow visitor mixins.
-"""
-
-from irx.builder.lowering.array import ArrayVisitorMixin, ArrowVisitorMixin
-
-__all__ = ["ArrayVisitorMixin", "ArrowVisitorMixin"]
