--- src/irx/builder/runtime/arrow/__init__.py
@@ -1,17 +1,0 @@
-"""
-title: Arrow-backed runtime feature compatibility exports for IRx.
-"""
-
-from irx.builder.runtime.arrow.feature import (
-    ARRAY_PRIMITIVE_TYPE_SPECS,
-    ARROW_PRIMITIVE_TYPE_SPECS,
-    build_array_runtime_feature,
-    build_arrow_runtime_feature,
-)
-
-__all__ = [
-    "ARRAY_PRIMITIVE_TYPE_SPECS",
-    "ARROW_PRIMITIVE_TYPE_SPECS",
-    "build_array_runtime_feature",
-    "build_arrow_runtime_feature",
-]
