+++ src/irx/builder/runtime/registry.py
@@ -9,10 +9,7 @@ from typing import TYPE_CHECKING, Iterable
 
 from llvmlite import ir
 
-from irx.builder.runtime.array.feature import (
-    build_array_runtime_feature,
-    build_arrow_runtime_feature,
-)
+from irx.builder.runtime.array.feature import build_array_runtime_feature
 from irx.builder.runtime.assertions.feature import (
     build_assertions_runtime_feature,
 )
@@ -300,5 +297,4 @@ def get_default_runtime_feature_registry() -> RuntimeFeatureRegistry:
     registry.register(build_libm_runtime_feature())
     registry.register(build_buffer_runtime_feature())
     registry.register(build_array_runtime_feature())
-    registry.register(build_arrow_runtime_feature())
     return registry
